procfs generator: move data to payloads
It seems more sensible to move the data types modeling the procfs
files for a process ID into the `lading_payloads` crate, so this
commit moves those types from the generator module of `lading` to the
`lading_payloads` crate.
